properly read header/footer & hide internals
Now there is no "magic" (e.g. not failing even if the size field reported a size that was too long).

Also, the old implementation had a Blocks() method. Since I discovered that there is a header instead of an initial, useless block, the meaning of Blocks() would have to change. This made me realize that I shouldn't expose anything internal to the format, since I don't know how the format works yet. It is best to expose a simple API that actually does what it says it does (e.g. get contents based on types), that way it can remain static no matter what I discover about the format.